+++ CHANGELOG.md
@@ -7,10 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED - DD.MM.YYYY]
 
+### Added
+
 ### Changed
 - updated publish-documentation workflow with two new steps
 - changed @ApiModelProperty annotation with @Schema
-
+- added logic to consume tombstone information from IRS for assets
+-
 ### Removed
 - removed frontend/dist folder from codeQL scan